fix(lint): split ui/routines/page.rs further under the 500-line gate
